# docs/api/openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "logaffe",
    "description": "The read and write surface of one logaffe installation. Every installation is at its own address.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/projects/{id}/entries": {
      "get": {
        "tags": [
          "logaffe"
        ],
        "summary": "One page of a project's entries, newest first by event time.",
        "operationId": "SearchEntries",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "From",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "Until",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "MinimumLevel",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Instance",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "LoggerName",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Trace",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Search",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Exception",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EntryPageResponse"
                }
              }
            }
          },
          "408": {
            "description": "Request Timeout",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReadExpiredResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpValidationProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/projects/{id}/entries/tail": {
      "get": {
        "tags": [
          "logaffe"
        ],
        "summary": "What has arrived since the last poll, in the view's order.",
        "operationId": "TailEntries",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "From",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "Until",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "MinimumLevel",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Instance",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "LoggerName",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Trace",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Search",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Exception",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "since",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TailResponse"
                }
              }
            }
          },
          "408": {
            "description": "Request Timeout",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReadExpiredResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpValidationProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/projects/{id}/entries/count": {
      "get": {
        "tags": [
          "logaffe"
        ],
        "summary": "How many entries a filter set matches, optionally grouped.",
        "operationId": "CountEntries",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "From",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "Until",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "MinimumLevel",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Instance",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "LoggerName",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Trace",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Search",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Exception",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "groupBy",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "bucket",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CountResponse"
                }
              }
            }
          },
          "408": {
            "description": "Request Timeout",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReadExpiredResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpValidationProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/projects/{id}/entries/{entryId}": {
      "get": {
        "tags": [
          "logaffe"
        ],
        "summary": "One entry in full, with its exception and its properties.",
        "operationId": "ReadEntry",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "entryId",
            "in": "path",
            "required": true,
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EntryResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CountResponse": {
        "required": [
          "groups"
        ],
        "type": "object",
        "properties": {
          "groups": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CountedGroupResponse"
            }
          }
        }
      },
      "CountedGroupResponse": {
        "required": [
          "value",
          "entries"
        ],
        "type": "object",
        "properties": {
          "value": {
            "type": [
              "null",
              "string"
            ]
          },
          "entries": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int64"
          }
        }
      },
...
      "EntryPageResponse": {
        "required": [
          "entries",
          "next"
        ],
        "type": "object",
        "properties": {
          "entries": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ListedEntryResponse"
            }
          },
          "next": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
      "EntryResponse": {
        "required": [
          "id",
          "eventTime",
          "receiptTime",
          "level",
          "loggerName",
          "instance",
          "trace",
          "span",
          "messageTemplate",
          "message",
          "exception",
          "properties",
          "messageTruncated",
          "exceptionTruncated"
        ],
        "type": "object",
        "properties": {
          "id": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int64"
          },
          "eventTime": {
            "type": "string",
            "format": "date-time"
          },
          "receiptTime": {
            "type": "string",
            "format": "date-time"
          },
          "level": {
            "type": "string"
          },
          "loggerName": {
            "type": [
              "null",
              "string"
            ]
          },
          "instance": {
            "type": [
              "null",
              "string"
            ]
          },
          "trace": {
            "type": [
              "null",
              "string"
            ]
          },
          "span": {
            "type": [
              "null",
              "string"
            ]
          },
          "messageTemplate": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "exception": {
            "type": [
              "null",
              "string"
            ]
          },
          "properties": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/JsonNode"
              }
            ]
          },
          "messageTruncated": {
            "type": "boolean"
          },
          "exceptionTruncated": {
            "type": "boolean"
          }
        }
      },
...
      "HttpValidationProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": [
              "null",
              "string"
            ]
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "detail": {
            "type": [
              "null",
              "string"
            ]
          },
          "instance": {
            "type": [
              "null",
              "string"
            ]
          },
          "errors": {
            "type": "object",
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        }
      },
...
      "JsonNode": {},
...
      "ListedEntryResponse": {
        "required": [
          "id",
          "eventTime",
          "level",
          "loggerName",
          "instance",
          "trace",
          "message",
          "messageTruncated",
          "hasException"
        ],
        "type": "object",
        "properties": {
          "id": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int64"
          },
          "eventTime": {
            "type": "string",
            "format": "date-time"
          },
          "level": {
            "type": "string"
          },
          "loggerName": {
            "type": [
              "null",
              "string"
            ]
          },
          "instance": {
            "type": [
              "null",
              "string"
            ]
          },
          "trace": {
            "type": [
              "null",
              "string"
            ]
          },
          "message": {
            "type": "string"
          },
          "messageTruncated": {
            "type": "boolean"
          },
          "hasException": {
            "type": "boolean"
          }
        }
      },
...
      "ReadExpiredResponse": {
        "required": [
          "narrow"
        ],
        "type": "object",
        "properties": {
          "narrow": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
...
      "TailResponse": {
        "required": [
          "entries",
          "next",
          "more"
        ],
        "type": "object",
        "properties": {
          "entries": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ListedEntryResponse"
            }
          },
          "next": {
            "type": "string"
          },
          "more": {
            "type": "boolean"
          }
        }
      },
...
    }
  },
...
}